Fold close-with-params demo into CSS MCP App demo
1. Extended demo/load-sidebar-v2-css-mcp-app/:
   - Added finish-and-close tool that calls closeAngieWithParams directly
   - Added "Finish and close" button in preview card
   - Added UI panel to display received params (safe DOM, no XSS)
   - Host tool calls closeAngieWithParams directly (not postMessage)

2. Deleted demo/load-sidebar-v2-close-with-params/ entirely

3. Updated README links:
   - Point to css-mcp-app for close-with-params demo
   - Removed references to deleted folder

4. SDK API and unit tests unchanged (no regressions)
   - All 269 tests pass
