--- demo/load-sidebar-v2-close-with-params/ajv-formats-shim.js
@@ -1,1 +1,0 @@
-export default function addFormats() {}
