--- demo/load-sidebar-v2-close-with-params/ajv-shim.js
@@ -1,15 +1,0 @@
-export default class Ajv {
-	compile() {
-		const validate = () => true;
-		validate.errors = null;
-		return validate;
-	}
-
-	getSchema() {
-		return undefined;
-	}
-
-	errorsText() {
-		return '';
-	}
-}
