+++ demo/load-sidebar-v2-css-mcp-app/preview-app.html
@@ -174,6 +174,7 @@
 		<div class="card-foot">
 			<span class="meta" id="meta">Waiting for the tool result…</span>
 			<button class="btn-ghost" id="discard" type="button" hidden>Discard</button>
+			<button class="btn-ghost" id="finish" type="button" hidden>Finish and close</button>
 			<button class="btn-primary" id="apply" type="button" hidden>Apply to page</button>
 		</div>
 	</div>
@@ -190,6 +191,7 @@
 	const metaEl = el( 'meta' );
 	const applyBtn = el( 'apply' );
 	const discardBtn = el( 'discard' );
+	const finishBtn = el( 'finish' );
 	const copyBtn = el( 'copy' );
 
 	/* ---------- JSON-RPC over postMessage ----------
@@ -346,6 +348,7 @@
 		setMeta( 'v' + ( proposal.version || 1 ) + ' · Proposed just now' );
 		applyBtn.hidden = false;
 		discardBtn.hidden = false;
+		finishBtn.hidden = false;
 		reportHeight();
 	}
 
@@ -398,8 +401,37 @@
 		settled = true;
 		applyBtn.hidden = true;
 		discardBtn.hidden = true;
+		finishBtn.hidden = true;
 		setMeta( 'Discarded · nothing was applied' );
 		reportHeight();
+	} );
+
+	finishBtn.addEventListener( 'click', async () => {
+		if ( settled ) {
+			return;
+		}
+
+		finishBtn.disabled = true;
+		applyBtn.disabled = true;
+		discardBtn.disabled = true;
+		setMeta( 'Finishing…' );
+
+		try {
+			await request( 'tools/call', {
+				name: 'finish-and-close',
+				arguments: { orderId: '456' },
+			} );
+
+			settled = true;
+			setMeta( 'Finished · Angie is closing', 'applied' );
+			reportHeight();
+		} catch ( error ) {
+			finishBtn.disabled = false;
+			applyBtn.disabled = false;
+			discardBtn.disabled = false;
+			setMeta( error.message || 'Could not finish', 'error' );
+			reportHeight();
+		}
 	} );
 
 	copyBtn.addEventListener( 'click', async () => {
